pyproject.toml
[build-system]
requires = [
  "setuptools>=62.1",
  "setuptools_scm[toml]>=6.2",
  "wheel",]
build-backend = "setuptools.build_meta"

[project]
name = "shone"
description = "Radiative transfer in JAX"
requires-python = ">=3.11"
readme = { file = "README.md", content-type = "text/x-rst" }
license = { file = "licenses/LICENSE.rst", content-type = "text/plain" }
authors = [
  { name = "Brett M. Morris, Jens Hoeijmakers", email = "morrisbrettm@gmail.com" },
]
dependencies = [
    "jax[cpu]",
    "jaxoplanet",
    "specutils",
    "periodictable",
    "dace-query",
    "xarray",
    "tensorflow-probability",
]
dynamic = ["version"]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-doctestplus",
    "pytest-cov",
    'pytest >=4.6.0',
    'pytest-astropy',
]
docs = [
    "sphinx",
    "sphinx-automodapi",
    "tomli; python_version <\"3.11\"",
    "sphinx-astropy[confv2]",
    "sphinx-toolbox",
    "matplotlib",
    "expecto",
    "sphinx-book-theme",
    "sphinx-github-style",
]

[project.urls]
repository = "https://github.com/bmorris3/shone"

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]

[tool.setuptools_scm]
write_to = "shone/version.py"

[tool.pytest.ini_options]
testpaths = [
  "shone",
  "docs",
]
doctest_plus = "enabled"
text_file_format = "rst"
addopts = "--doctest-rst"

[tool.coverage.run]
omit = [
 "shone/__init*",
 "shone/conftest.py",
 "shone/*setup_package*",
 "shone/tests/*",
 "shone/*/tests/*",
 "shone/extern/*",
 "shone/version*",
 "*/shone/__init*",
 "*/shone/conftest.py",
 "*/shone/*setup_package*",
 "*/shone/tests/*",
 "*/shone/*/tests/*",
 "*/shone/extern/*",
 "*/shone/version*",
]

[tool.coverage.report]
exclude_lines = [
  # Have to re-enable the standard pragma
  "pragma: no cover",
  # Don't complain about packages we have installed
  "except ImportError",
  # Don't complain if tests don't hit assertions
  "raise AssertionError",
  "raise NotImplementedError",
  # Don't complain about script hooks
  "def main(.*):",
  # Ignore branches that don't pertain to this version of Python
  "pragma: py{ignore_python_version}",
  # Don't complain about IPython completion helper
  "def _ipython_key_completions_",
  # typing.TYPE_CHECKING is False at runtime
  "if TYPE_CHECKING:",
  # Ignore typing overloads
  "@overload",
]
